Fix Txn state handling on failure in #17 and #19
This commit includes a test for the map full case and modifications to
the Txn state handling.

The fault was caused because a failed Txn commit (caused by the map
being full) would mean the Txn handle becomes invalid, and the subsequent
Txn.close() call to mdb_txn_abort is be invalid for an already-closed
handle.