gh-185 Add protection for closed env
Change Cursor and Txn to hold the Env instance so it can be checked
for its open/closed state.
Add calls to env.checkNotClosed() in methods that call out to the LMDB
lib to prevent the JVM crashing if the env has been closed.
Add lots of tests involving env closure.