Update GitHub Actions Workflow to modern actions and versions
Support for dependencies in the Sonatype snapshot repository has been
removed because the maven-settings-action does not presently support
editing the settings.xml produced by the setup-java action. The only
historical reason to add the Sonatype snapshot repository was to test
with the Acegi Standard Project superpom snapshots. This is usually
undertaken on a developer machine and then an Acegi Standard Project
superpom is released to Maven Central. As such having GitHub Actions
able to perform this testing is largely unnecessary and has been
removed for maintenance practicality.