Update formatting, documentation and limit GC test write frequency
The GarbageCollectionTest performing 1,000 GC operations caused a
substantial increase in test execution time. Trial and error found
commenting out the new ReferenceUtil.reachabilityFence0 references
resulted in the failures being consistently detected with a value of
250. Nevertheless even a value of 250 would result in the new test
being approximately 3 times longer in duration than any other test. To
resolve this the default value was reduced to 50 and a system property
introduced to allow any arbitrary value to be used. In turn the GitHub
Action was amended to test with 1,000 GC-inducing writes as per the
original code.

As an aside on OpenJDK Java 20 the mere presence of a call to
ReferenceUtil.reachabilityFence0 was sufficient to pass the
tests (even with high values such as 2,000) despite there being no
method body remaining in reachabilityFence0. As such it is unverified
whether the synchronization is actually required for the strong
reachability. This is consistent with
https://github.com/apache/datasketches-memory/issues/91#issuecomment-444294778.
In any event synchronization has been retained given it is the more
conservative approach.